# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是一個 Menu2What 的 API 服務",
        "title": "Menu2What API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6382",
    "basePath": "/",
    "paths": {
        "/gemini/api/image_and_text": {
            "post": {
                "description": "上傳圖片和文字到 Gemini AI 並獲取分析回應",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gemini"
                ],
                "summary": "Gemini AI 圖文分析 API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文字訊息",
                        "name": "message",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "圖片檔案",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.GeminiAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.GeminiAPIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.GeminiAPIErrorResponse"
                        }
                    }
                }
            }
        },
        "/gemini/api/text_only": {
            "post": {
                "description": "傳送訊息到 Gemini AI 並獲取回應",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gemini"
                ],
                "summary": "Gemini AI API",
                "parameters": [
                    {
                        "description": "請求內容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.GeminiAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.GeminiAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.GeminiAPIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.GeminiAPIErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.GeminiAPIErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "無效的輸入格式"
                },
                "error": {
                    "type": "string",
                    "example": "處理請求時發生錯誤"
                }
            }
        },
        "controllers.GeminiAPIRequest": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "你好"
                }
            }
        },
        "controllers.GeminiAPIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "你好，我是 Gemini AI"
                }
            }
        }
    }
}